Added `MainActor` & `Sendable` Annotations (#101)
This PR adds the `@MainActor` annotation, and thus `MainActor` isolation, to all relevant types and closures in SectionKit and DiffingSectionKit.
This makes it possible for the compiler to statically detect, whether a function might be called from a background thread.
Since SectionKit interacts with `UICollectionView`, it is very important, that all accesses to its functions are performed from the main thread.

Theoretically, using the `MainActor` class requires iOS 13, but I've tested the annotation with a project that targets iOS 11 (the oldest iOS version that Xcode 14 supports) and it compiled without complaints (it seems the actor isolation is toolchain specific, whereas calling a function like `MainActor.run` is iOS version specific). This means, the only requirement is Swift 5.5 (which is shipped with Xcode 13.0). Thus this PR increases the `swift-tools-version` to 5.5 to fit that requirement.

Additionally, some types have been marked as `Sendable`.

**Please note**: This is potentially a source breaking change.

Signed-off-by: Niclas Kristek <dev@nkristek.com>